refactor: rename Public..Post to General..Post
