+++ web-client/src/ducks/GeneralOffers/reducer.ts
