+++ web-client/src/ducks/GeneralOffers/types.ts
