+++ web-client/src/ducks/ResponseOffers/actions.ts
@@ -15,7 +15,7 @@ export const dispatchCreateResponseOfferFromRequest = (
 ) => (dispatch: Function) => {
   const postPayload = Post.factory(payload);
   postPayload.requestingHelp = !payload.requestingHelp;
-  postPayload.sourcePublicPostId = payload.postId;
+  postPayload.sourceGeneralPostId = payload.postId;
   postPayload.isResponse = true;
   return dispatch({
     type: DISPATCH_CREATE_PRIVATE_OFFER,
